Up and running on new Marked API
Switched over to the Marked plugin model that is now available so we don't have to manage it ourselves. Prior status restored and the system should be a lot easier to expand now.

With this change, I had to switch the way menus are built. All the pages markedown is now rendered before we do anything else, and the menu is built dynamically as part of that process. That likely means that issue #13 which I had a fix pending for will need some more thought. Another minor change is that IDs are now unique throughout the documentation. Should be minor as it would only effect duplicated IDs from different pages, which was broken in the menu anyway.