Convert links and pull in assets
Automatically converts links to other pages to point to the right .html file. Also pulls in any locally linked assets and includes them in a new "assets" folder in the export.

Need to put this and the images one behind a flag. May be cases when you don't want this behavior.